Fix ReflectiveConnectionMonitoringTransformer
Moved Dispatcher to java.lang package so it can be loaded inside JBoss
