+++ build.gradle
@@ -53,6 +53,7 @@ ext {
 	slf4jVersion = '1.7.18'
 	jacksonVersion = '2.7.2'
 	esVersion = '2.2.0'
+	byteBuddyVersion = '1.3.13'
 }
 
 subprojects {
