+++ stagemonitor-core/build.gradle
@@ -6,8 +6,9 @@ dependencies {
 	compile "io.dropwizard.metrics:metrics-json:$metricsVersion"
 	compile "com.fasterxml.jackson.core:jackson-databind:$jacksonVersion"
 	compile "com.fasterxml.jackson.module:jackson-module-afterburner:$jacksonVersion"
-	compile "net.bytebuddy:byte-buddy:1.3.12"
-	compile "net.bytebuddy:byte-buddy-agent:1.3.12"
+	compile "net.bytebuddy:byte-buddy:$byteBuddyVersion"
+	compile "net.bytebuddy:byte-buddy-agent:$byteBuddyVersion"
+	provided project(":stagemonitor-dispatcher")
 
 	testCompile "org.elasticsearch:elasticsearch:$esVersion"
 	testCompile 'commons-io:commons-io:2.4'
