+++ stagemonitor-core/src/main/java/org/stagemonitor/core/instrument/AgentAttacher.java
@@ -8,8 +8,13 @@ import static net.bytebuddy.matcher.ElementMatchers.not;
 import static org.stagemonitor.core.instrument.ReflectionClassLoaderMatcher.isReflectionClassLoader;
 import static org.stagemonitor.core.instrument.TimedElementMatcherDecorator.timed;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.lang.instrument.ClassFileTransformer;
 import java.lang.instrument.Instrumentation;
+import java.lang.stagemonitor.dispatcher.Dispatcher;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
@@ -17,6 +22,7 @@ import java.util.List;
 import java.util.ServiceLoader;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.jar.JarFile;
 
 import net.bytebuddy.ByteBuddy;
 import net.bytebuddy.agent.ByteBuddyAgent;
@@ -28,18 +34,20 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.stagemonitor.core.CorePlugin;
 import org.stagemonitor.core.Stagemonitor;
+import org.stagemonitor.core.util.IOUtils;
 
 /**
  * Attaches the {@link ByteBuddyAgent} at runtime and registers all {@link StagemonitorByteBuddyTransformer}s
  */
 public class AgentAttacher {
 
+	private static final Logger logger = LoggerFactory.getLogger(AgentAttacher.class);
+	private static final String DISPATCHER_CLASS_NAME = "java.lang.stagemonitor.dispatcher.Dispatcher";
+	private static final String IGNORED_CLASSLOADERS_KEY = AgentAttacher.class.getName() + "hashCodesOfClassLoadersToIgnore";
 	private static final Runnable NOOP_ON_SHUTDOWN_ACTION = new Runnable() {
 		public void run() {
 		}
 	};
-	private static final Logger logger = LoggerFactory.getLogger(AgentAttacher.class);
-	private static final String IGNORED_CLASSLOADERS_KEY = AgentAttacher.class.getName() + "hashCodesOfClassLoadersToIgnore";
 
 	private static CorePlugin corePlugin = Stagemonitor.getPlugin(CorePlugin.class);
 	private static boolean runtimeAttached = false;
@@ -50,8 +58,9 @@ public class AgentAttacher {
 			.or(nameStartsWith("com.sun."))
 			.or(nameStartsWith("sun."))
 			.or(nameStartsWith("jdk."))
-			.or(nameStartsWith("org.aspectj"))
-			.or(nameStartsWith("org.groovy"))
+			.or(nameStartsWith("org.aspectj."))
+			.or(nameStartsWith("org.groovy."))
+			.or(nameStartsWith("net.bytebuddy."))
 			.or(nameContains("javassist"))
 			.or(nameContains(".asm."));
 
@@ -98,7 +107,7 @@ public class AgentAttacher {
 			} catch (IllegalStateException e) {
 				instrumentation = ByteBuddyAgent.install();
 			}
-			Dispatcher.init(instrumentation);
+			ensureDispatcherIsAppendedToBootstrapClasspath(instrumentation);
 			if (!Dispatcher.getValues().containsKey(IGNORED_CLASSLOADERS_KEY)) {
 				Dispatcher.put(IGNORED_CLASSLOADERS_KEY, Collections.newSetFromMap(new ConcurrentHashMap<Integer, Boolean>()));
 			}
@@ -109,6 +118,28 @@ public class AgentAttacher {
 					"You can try loadint the agent with the command line argument -javaagent:/path/to/byte-buddy-agent-<version>.jar", e);
 			return false;
 		}
+	}
+
+	private static Class<?> ensureDispatcherIsAppendedToBootstrapClasspath(Instrumentation instrumentation)
+			throws ClassNotFoundException, IOException {
+		final ClassLoader bootstrapClassloader = ClassLoader.getSystemClassLoader().getParent();
+		try {
+			return bootstrapClassloader.loadClass(DISPATCHER_CLASS_NAME);
+			// already injected
+		} catch (ClassNotFoundException e) {
+			final JarFile jarfile = new JarFile(createTempDispatcherJar());
+			instrumentation.appendToBootstrapClassLoaderSearch(jarfile);
+			return bootstrapClassloader.loadClass(DISPATCHER_CLASS_NAME);
+		}
+	}
+
+	private static File createTempDispatcherJar() throws IOException {
+		final InputStream input = AgentAttacher.class.getClassLoader()
+				.getResourceAsStream("stagemonitor-dispatcher.jar.gradlePleaseDontExtract");
+		final File tempDispatcherJar = File.createTempFile("stagemonitor-dispatcher", ".jar");
+		tempDispatcherJar.deleteOnExit();
+		IOUtils.copy(input, new FileOutputStream(tempDispatcherJar));
+		return tempDispatcherJar;
 	}
 
 	private static List<ClassFileTransformer> initByteBuddyClassFileTransformers() {
