+++ stagemonitor-jdbc/build.gradle
@@ -2,7 +2,6 @@ dependencies {
 	compile project(":stagemonitor-requestmonitor")
 	compile 'p6spy:p6spy:2.2.0'
 
-	provided project(":stagemonitor-dispatcher")
 	testCompile 'org.apache.tomcat:tomcat-jdbc:8.5.0'
 	testCompile 'org.hsqldb:hsqldb:2.3.3'
 }
